Re-architect into three layers
