+++ bluetti_mqtt/bluetooth/__init__.py
@@ -1,106 +1,15 @@
-import asyncio
 import logging
 import re
-import time
-from typing import Dict, List, Set
-from bleak import BleakError, BleakScanner
+from typing import Set
+from bleak import BleakScanner
 from bleak.backends.device import BLEDevice
-from bluetti_mqtt.bus import CommandMessage, EventBus, ParserMessage
-from bluetti_mqtt.commands import QueryRangeCommand
-from bluetti_mqtt.devices import BluettiDevice, AC200M, AC300, EP500, EP500P, EB3A
-from bluetti_mqtt.bluetooth.client import BluetoothClient
-from bluetti_mqtt.bluetooth.exc import BadConnectionError, InvalidRequestError, ParseError
+from bluetti_mqtt.core import BluettiDevice, AC200M, AC300, EP500, EP500P, EB3A
+from .client import BluetoothClient
+from .exc import BadConnectionError, InvalidRequestError, ParseError
+from .manager import MultiDeviceManager
 
 
 DEVICE_NAME_RE = re.compile(r'^(AC200M|AC300|EP500P|EP500|EB3A)(\d+)$')
-
-
-class BluetoothClientHandler:
-    clients: Dict[BluettiDevice, BluetoothClient]
-
-    def __init__(self, devices: List[BluettiDevice], interval: int, bus: EventBus):
-        self.devices = devices
-        self.interval = interval
-        self.bus = bus
-        self.clients = {}
-
-    async def run(self):
-        loop = asyncio.get_running_loop()
-
-        # Start clients
-        addresses = [d.address for d in self.devices]
-        logging.info(f'Connecting to clients: {addresses}')
-        self.clients = {d: BluetoothClient(d.address) for d in self.devices}
-        self.client_tasks = [loop.create_task(c.run()) for c in self.clients.values()]
-
-        # Connect to event bus
-        self.bus.add_command_listener(self.handle_command)
-
-        # Poll the clients
-        logging.info('Starting to poll clients...')
-        polling_tasks = [self._poll(d, c) for d, c in self.clients.items()]
-        pack_polling_tasks = [self._pack_poll(d, c) for d, c in self.clients.items() if len(d.pack_logging_commands) > 0]
-        await asyncio.gather(*(polling_tasks + pack_polling_tasks))
-
-    async def handle_command(self, msg: CommandMessage):
-        if msg.device in self.clients:
-            client = self.clients[msg.device]
-            logging.debug(f'Performing command {msg.device}: {msg.command}')
-            await client.perform_nowait(msg.command)
-
-    async def _poll(self, device: BluettiDevice, client: BluetoothClient):
-        while True:
-            if not client.is_connected:
-                logging.debug(f'Waiting for connection to {device.address} to start polling...')
-                await asyncio.sleep(1)
-                continue
-
-            # Send all polling commands
-            start_time = time.monotonic()
-            for command in device.polling_commands:
-                await self._poll_with_command(device, client, command)
-            elapsed = time.monotonic() - start_time
-
-            # Limit polling rate if interval provided
-            if self.interval > 0 and self.interval > elapsed:
-                await asyncio.sleep(self.interval - elapsed)
-
-    async def _pack_poll(self, device: BluettiDevice, client: BluetoothClient):
-        while True:
-            if not client.is_connected:
-                logging.debug(f'Waiting for connection to {device.address} to start pack polling...')
-                await asyncio.sleep(1)
-                continue
-
-            start_time = time.monotonic()
-            for pack in range(1, device.pack_num_max + 1):
-                # Send pack set command if the device supports more than 1 pack
-                if device.pack_num_max > 1:
-                    command = device.build_setter_command('pack_num', pack)
-                    await client.perform_nowait(command)
-                    await asyncio.sleep(10) # We need to wait after switching packs for the data to be available
-
-                # Poll
-                for command in device.pack_logging_commands:
-                    await self._poll_with_command(device, client, command)
-            elapsed = time.monotonic() - start_time
-
-            # Limit polling rate if interval provided
-            if self.interval > 0 and self.interval > elapsed:
-                await asyncio.sleep(self.interval - elapsed)
-
-    async def _poll_with_command(self, device: BluettiDevice, client: BluetoothClient, command: QueryRangeCommand):
-        result_future = await client.perform(command)
-        try:
-            result = await result_future
-            parsed = device.parse(command.page, command.offset, result[3:-2])
-            await self.bus.put(ParserMessage(device, parsed))
-        except ParseError:
-            logging.debug('Got a parse exception...')
-        except InvalidRequestError:
-            logging.debug(f'Got an invalid request error for {command}: {err}')
-        except (BadConnectionError, BleakError) as err:
-            logging.debug(f'Needed to disconnect due to error: {err}')
 
 
 async def scan_devices():
