+++ bluetti_mqtt/bus.py
@@ -2,8 +2,7 @@ import asyncio
 from dataclasses import dataclass
 import logging
 from typing import Callable, List, Union
-from bluetti_mqtt.commands import DeviceCommand
-from bluetti_mqtt.devices import BluettiDevice
+from bluetti_mqtt.core import BluettiDevice, DeviceCommand
 
 @dataclass(frozen=True)
 class ParserMessage:
