+++ bluetti_mqtt/core/devices/ac200m.py
@@ -1,6 +1,6 @@
 from enum import Enum, unique
 from typing import List
-from bluetti_mqtt.commands import QueryRangeCommand
+from ..commands import QueryRangeCommand
 from .bluetti_device import BluettiDevice
 from .struct import DeviceStruct
 
@@ -15,14 +15,6 @@ class OutputMode(Enum):
 
 
 @unique
-class UpsMode(Enum):
-    CUSTOMIZED = 1
-    PV_PRIORITY = 2
-    STANDARD = 3
-    TIME_CONTROL = 4
-
-
-@unique
 class AutoSleepMode(Enum):
     THIRTY_SECONDS = 2
     ONE_MINUTE = 3
@@ -30,7 +22,7 @@ class AutoSleepMode(Enum):
     NEVER = 5
 
 
-class EP500P(BluettiDevice):
+class AC200M(BluettiDevice):
     def __init__(self, address: str, sn: str):
         self.struct = DeviceStruct()
 
@@ -50,42 +42,33 @@ class EP500P(BluettiDevice):
 
         # Page 0x00 - Details
         self.struct.add_enum_field('ac_output_mode', 0x00, 0x46, OutputMode)
-        self.struct.add_decimal_field('internal_ac_voltage', 0x00, 0x47, 1)
+        self.struct.add_uint_field('internal_ac_voltage', 0x00, 0x47)
         self.struct.add_decimal_field('internal_current_one', 0x00, 0x48, 1)
         self.struct.add_uint_field('internal_power_one', 0x00, 0x49)
-        self.struct.add_decimal_field('internal_ac_frequency', 0x00, 0x4A, 2)
-        self.struct.add_decimal_field('internal_current_two', 0x00, 0x4B, 1)
-        self.struct.add_uint_field('internal_power_two', 0x00, 0x4C)
-        self.struct.add_decimal_field('ac_input_voltage', 0x00, 0x4D, 1)
-        self.struct.add_decimal_field('internal_current_three', 0x00, 0x4E, 1)
-        self.struct.add_uint_field('internal_power_three', 0x00, 0x4F)
-        self.struct.add_decimal_field('ac_input_frequency', 0x00, 0x50, 2)
+        self.struct.add_decimal_field('internal_ac_frequency', 0x00, 0x4A, 1)
         self.struct.add_decimal_field('dc_input_voltage', 0x00, 0x56, 1)
         self.struct.add_uint_field('dc_input_power', 0x00, 0x57)
         self.struct.add_decimal_field('dc_input_current', 0x00, 0x58, 1)
 
         # Page 0x00 - Battery Data
         self.struct.add_uint_field('pack_num_max', 0x00, 0x5B)
-        self.struct.add_decimal_field('pack_voltage', 0x00, 0x5C, 1) # Full pack voltage
-        self.struct.add_uint_field('pack_battery_percent', 0x00, 0x5E)
         self.struct.add_uint_field('pack_num', 0x00, 0x60)
+        self.struct.add_decimal_field('pack_voltage', 0x00, 0x62, 2) # Full pack voltage
+        self.struct.add_uint_field('pack_battery_percent', 0x00, 0x63)
         self.struct.add_decimal_array_field('cell_voltages', 0x00, 0x69, 16, 2)
 
         # Page 0x0B - Controls
-        self.struct.add_enum_field('ups_mode', 0x0B, 0xB9, UpsMode)
         self.struct.add_uint_field('pack_num', 0x0B, 0xBE)
         self.struct.add_bool_field('ac_output_on', 0x0B, 0xBF)
         self.struct.add_bool_field('dc_output_on', 0x0B, 0xC0)
-        self.struct.add_bool_field('grid_charge_on', 0x0B, 0xC3)
-        self.struct.add_bool_field('time_control_on', 0x0B, 0xC5)
-        self.struct.add_uint_field('battery_range_start', 0x0B, 0xC7)
-        self.struct.add_uint_field('battery_range_end', 0x0B, 0xC8)
         # 0xD7-0xD9 is the current device time & date without a timezone
-        self.struct.add_bool_field('bluetooth_connected', 0x0B, 0xDC)
-        # 0xDF-0xF0 is the time control programming
         self.struct.add_enum_field('auto_sleep_mode', 0x0B, 0xF5, AutoSleepMode)
 
-        super().__init__(address, 'EP500P', sn)
+        super().__init__(address, 'AC200M', sn)
+
+    @property
+    def pack_num_max(self):
+        return 3
 
     @property
     def polling_commands(self) -> List[QueryRangeCommand]:
