+++ bluetti_mqtt/core/devices/struct.py
